+++ src/LuaPrinter.ts
@@ -1,11 +1,13 @@
+import * as path from "path";
 import { Mapping, SourceMapGenerator, SourceNode } from "source-map";
+import { getEmitPath } from ".";
 import * as ts from "typescript";
 import { CompilerOptions, isBundleEnabled, LuaLibImportKind } from "./CompilerOptions";
 import * as lua from "./LuaAST";
 import { loadLuaLibFeatures, LuaLibFeature } from "./LuaLib";
 import { isValidLuaIdentifier } from "./transformation/utils/safe-names";
 import { EmitHost } from "./transpilation";
-import { intersperse, trimExtension } from "./utils";
+import { intersperse, normalizeSlashes } from "./utils";
 
 // https://www.lua.org/pil/2.4.html
 // https://www.ecma-international.org/ecma-262/10.0/index.html#table-34
@@ -120,14 +122,17 @@ export class LuaPrinter {
     };
 
     private currentIndent = "";
-    private sourceFile: string;
+    private luaFile: string;
+    private relativeSourcePath: string;
     private options: CompilerOptions;
 
     public static readonly sourceMapTracebackPlaceholder = "{#SourceMapTraceback}";
 
-    constructor(private emitHost: EmitHost, program: ts.Program, fileName: string) {
+    constructor(private emitHost: EmitHost, private program: ts.Program, private sourceFile: string) {
         this.options = program.getCompilerOptions();
-        this.sourceFile = fileName;
+        this.luaFile = normalizeSlashes(getEmitPath(this.sourceFile, this.program));
+        // Source nodes contain relative path from mapped lua file to original TS source file
+        this.relativeSourcePath = normalizeSlashes(path.relative(path.dirname(this.luaFile), this.sourceFile));
     }
 
     public print(file: lua.File): PrintResult {
@@ -230,12 +235,12 @@ export class LuaPrinter {
         const { line, column } = lua.getOriginalPos(node);
 
         return line !== undefined && column !== undefined
-            ? new SourceNode(line + 1, column, this.sourceFile, chunks, name)
-            : new SourceNode(null, null, this.sourceFile, chunks, name);
+            ? new SourceNode(line + 1, column, this.relativeSourcePath, chunks, name)
+            : new SourceNode(null, null, this.relativeSourcePath, chunks, name);
     }
 
     protected concatNodes(...chunks: SourceChunk[]): SourceNode {
-        return new SourceNode(null, null, this.sourceFile, chunks);
+        return new SourceNode(null, null, this.relativeSourcePath, chunks);
     }
 
     protected printBlock(block: lua.Block): SourceNode {
@@ -757,7 +762,7 @@ export class LuaPrinter {
     }
 
     public printOperator(kind: lua.Operator): SourceNode {
-        return new SourceNode(null, null, this.sourceFile, LuaPrinter.operatorMap[kind]);
+        return new SourceNode(null, null, this.relativeSourcePath, LuaPrinter.operatorMap[kind]);
     }
 
     protected joinChunksWithComma(chunks: SourceChunk[]): SourceChunk[] {
@@ -787,7 +792,7 @@ export class LuaPrinter {
     // will not generate 'empty' mappings in the source map that point to nothing in the original TS.
     private buildSourceMap(sourceRoot: string, rootSourceNode: SourceNode): SourceMapGenerator {
         const map = new SourceMapGenerator({
-            file: trimExtension(this.sourceFile) + ".lua",
+            file: path.basename(this.luaFile),
             sourceRoot,
         });