+++ test/util.ts
@@ -65,6 +65,19 @@ export function testEachVersion<T extends TestBuilder>(
     }
 }
 
+export function expectEachVersionExceptJit<T>(
+    expectation: (builder: T) => void
+): Record<tstl.LuaTarget, ((builder: T) => void) | boolean> {
+    return {
+        [tstl.LuaTarget.Universal]: expectation,
+        [tstl.LuaTarget.Lua51]: expectation,
+        [tstl.LuaTarget.Lua52]: expectation,
+        [tstl.LuaTarget.Lua53]: expectation,
+        [tstl.LuaTarget.Lua54]: expectation,
+        [tstl.LuaTarget.LuaJIT]: false, // Exclude JIT
+    };
+}
+
 const memoize: MethodDecorator = (_target, _propertyKey, descriptor) => {
     const originalFunction = descriptor.value as any;
     const memoized = new WeakMap();
@@ -491,6 +504,8 @@ end)());`;
         try {
             result = vm.runInContext(this.getJsCodeWithWrapper(), globalContext);
         } catch (error) {
+            const hasMessage = (error: any): error is { message: string } => error.message !== undefined;
+            assert(hasMessage(error));
             return new ExecutionError(error.message);
         }